refactor(auth): formalize the auth module boundary and move login/register
