+++ src/app/index.ts
@@ -1,2 +1,2 @@
 export { AppProviders } from './providers'
-export { AppRoutes, GuestOnlyRoute, ProtectedRoute } from './routes'
+export { AppRoutes } from './routes'
