+++ src/app/routes/index.ts
@@ -1,2 +1,1 @@
 export { AppRoutes } from './AppRoutes'
-export { GuestOnlyRoute, ProtectedRoute } from './RouteGuards'
