--- src/auth/index.ts
@@ -1,2 +1,0 @@
-export { AUTH_SESSION_STORAGE_KEY, AuthProvider, useAuth } from './AuthContext'
-export type { AuthenticatedUser } from './AuthContext'
