+++ src/modules/auth/authContext.tsx
@@ -1,0 +1,14 @@
+import { createContext, useContext } from 'react'
+import type { AuthContextValue } from './types'
+
+export const AuthContext = createContext<AuthContextValue | undefined>(undefined)
+
+export function useAuth() {
+  const context = useContext(AuthContext)
+
+  if (!context) {
+    throw new Error('useAuth must be used within an AuthProvider.')
+  }
+
+  return context
+}
