+++ src/modules/my-scenes/MyScenesPage.test.tsx
@@ -3,9 +3,9 @@ import userEvent from '@testing-library/user-event'
 import { MemoryRouter, Route, Routes } from 'react-router-dom'
 import { describe, expect, it, vi } from 'vitest'
 import { AUTH_SESSION_STORAGE_KEY, AuthProvider, type AuthenticatedUser } from '@auth'
+import { LoginPage } from '@modules/auth'
 import { buildApiUrl } from '@lib/api'
 import { SceneDetailPage } from '@modules/scene-detail'
-import { LoginPage } from '@pages/LoginPage'
 import { MyScenesPage } from './MyScenesPage'
 
 vi.mock('@modules/player', async (importOriginal) => {
