+++ src/pages/index.ts
@@ -1,9 +1,8 @@
 export { CreateScenePage } from './CreateScenePage'
 export { HomePage } from './HomePage'
-export { LoginPage } from './LoginPage'
+export { LoginPage, RegisterPage } from '@modules/auth'
 export { MyScenesPage } from '@modules/my-scenes'
 export { PlaceholderPage } from './PlaceholderPage'
-export { RegisterPage } from './RegisterPage'
 export { SceneDetailPage } from '@modules/scene-detail'
 export { ScenesPage } from '@modules/discovery'
 export { SettingsPage } from './SettingsPage'
