+++ vite.config.ts
@@ -11,7 +11,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@app': fileURLToPath(new URL('./src/app', import.meta.url)),
-      '@auth': fileURLToPath(new URL('./src/auth', import.meta.url)),
+      '@auth': fileURLToPath(new URL('./src/modules/auth', import.meta.url)),
       '@components': fileURLToPath(new URL('./src/components', import.meta.url)),
       '@lib': fileURLToPath(new URL('./src/lib', import.meta.url)),
       '@modules': fileURLToPath(new URL('./src/modules', import.meta.url)),
